Time: 607 ms (5.01%), Space: 192.5 MB (5.12%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -16,7 +16,7 @@ public:
         prefix[0] = 1;
         return dfs(root, 0, targetSum, prefix);
     }
-    int dfs(TreeNode* root, int currsum, int target, unordered_map<long long, int> prefix){
+    int dfs(TreeNode* root, long long currsum, int target, unordered_map<long long, int> prefix){
         if(!root) return 0;
         currsum+= root->val;
         int count = 0;
